requirements.txt
transformers
diffusers
torch
accelerate
gradio
Pillow
